Ignore possible IO error on buffer write for subscribe and unsubscribe protocols.
For those two protocols, we write to the buffer when we know we are not in the process of reconnecting (using the pending buffer). Still, there could be a case where the connection has just been lost and the write to the buffer would cause this buffer to be flushed (causing a socket write), which would result in an error. However, if we successfully return the subscription object to the user and store it internally, we are fine since after the reconnect, the library sends all subscriptions, so we are ok.