Fix multiple natsSubscription_NextMsg on same sub could block.
In the case of an asynchronous subscription, there was only one thread
to notify (when a message is received or when the subscription is
closed). However, in the case of sync subscriptions, it is possible that
the user calls NextMsg on the same subscription from several threads.
Therefore, notification should have been Broadcast instead of Signal.
Also, indication that we are in a wait has been changed from a boolean
to an int.