Fixed reconnect loop possibly not yielding to natsConnection_Close()
The issue manifests on Windows with a ReconnectWait lower than a
second, because it takes more than a second for createConn() to
return when no server is running. Therefore, the loop would never
sleep (and never release the lock), which would prevent a
natsConnection_Close() from breaking the reconnect loop.