[ADDED] Ability to create asynchronous subscriptions with timeout
Added two new APIs

- `natsConnection_SubscribeTimeout`
- `natsConnection_QueueSubscribeTimeout`

This APIs require a timeout value. When no message is received
after the timeout interval elapses, the regular eessage handler
is invoked but with `NULL` message to indicate that this is a
timeout condition.
The subscription can be destroyed then, otherwise, the callback
will fire again when messages are received or a timeout occurs
again.

An example to demonstrate the use of those APIs has been added.
Check `examples/asynctimeout.c`.