[ADDED] Ability to process asynchronous INFO protocol
NATS Server 0.9.2+ are now sending async INFO protocols to clients
that support it (clients now send a protocol version in the
CONNECT protocol). This INFO protocol may contain an array of
URLs of servers in the mesh. This will be used by the client
library during reconnect.

Added two options setters:
- natsOptions_SetUserInfo
- natsOptions_SetToken
These may be required to have the client library able to reconnect
to servers that the library got from asynchronous INFO sent by
servers 0.9.2+