[FIXED] Permissions violation handling
If the application publishes or subscribes to a subject for which
the user does not have proper permissions, the client library
receives an error from the server. This needs to be handled properly
otherwise the connection ends-up being closed.
The library will now invoke the error handler if one is set.