[FIXED] Possible crash after a reconnect
In the reconnect thread, the connection's field reconnectThread
was set to NULL, lock released in order to flush.
If at that time the connection was closed, since reconnecThread
was NULL, the thread doing the close will not join the reconnect
thread and release the connection. When the reconnect thread
would try to flush, it would possibly crash since the connection
object was being freed.

Resolves #77