[FIXED] Installed adapters headers
Modified adapters (libevent and libuv) headers to include
"../nats.h". This will work even in case when we build examples
with cmake that includes `src`, and also for users using
`/usr/local/include` and having their files include `nats/nats.h`
and `nats/adapters/libuv.h` (for instance).

Updated libuv adapter to use uv_poll_init() instead of uv_poll_init_socket()
since the later disappeared from libuv's master repo.

For non Windows builds, if examples are built statically, link
against static libevent/libuv libraries.
For Windows, if say "libuv.lib" is the lib from the static built of
that library, then the example will be statically linked with that.
If the ".lib" file is from the building of a shared library, the
example will need "libuv.dll" to be able to run.

On Windows, changed the "libevent" library name to "event" since
it looks like this is how the library is named now.

Resolves #90