[FIXED] Installed adapters headers
Adapters (libevent and libuv) installed headers (located in src/adapters)
need to include "nats/nats.h" since cnats headers are installed in
{install_prefix}/include/nats.

However, in order to be able to build the examples without having
to first build and install the library (without building the
examples), I have a copy of the adapters directory in the
examples directory. These version include "nats.h", instead of
"nats/nats.h". I don't like that too much, but otherwise, the
only alternative I see would be to default to not build examples,
and force user to do a make install, then make edit_cache to
turn ON building of examples and build again.

Updated libuv adapter to use uv_poll_init() instead of uv_poll_init_socket()
since the later disappeared from libuv's master repo.

For non Windows builds, if examples are built statically, link
against static libevent/libuv libraries.
For Windows, if say "libuv.lib" is the lib from the static built of
that library, then the example will be statically linked with that.
If the ".lib" file is from the building of a shared library, the
example will need "libuv.dll" to be able to run.

On Windows, changed the "libevent" library name to "event" since
it looks like this is how the library is named now.

Resolves #90