[ADDED] NoEcho feature
A new option can be set with `natsOptions_SetNoEcho()` to instruct
the server to not send back messages originating from a connection
to its matching subscriptions.

This requires server to be at version 1.2.0 and above. If the
option is set to true and connection is maded agains an older
server, the connect will fail with error NATS_NO_SERVER_SUPPORT.

Resolves #145

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>