[ADDED] Ability to get underlying NATS Connection
The function `stanConnection_GetNATSConnection()` will return
the underlying NATS Connection, which cannot be explicitly closed
or destroyed.
The connection is still managed and owned by the streaming connection.

This call is to be able to do non-streaming pub/sub without having
to create a separate NATS Connection.

Related to PR #166 from @edam
Resolves #165

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>