[CHANGED] natsConnStatus enum values have now a prefix
To avoid name collisions when used with other libraries that
also defined these common terms. The connection status enum
values have been prefixed with `NATS_CONN_STATUS_` prefix.

A compilation option has also been added to allow the use of
the original enum values in case user does not have name collisions
and referenced values in the application and not willing to
update those. The option is disabled by default.

Resolves #168

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>