[FIXED] Possible core on process exit
We attach a destructor on process exit to do final cleanup of
the library. However, some NATS threads may have still be running
and accessing locks/etc.. that were freed. Make sure to retain/release
the library in NATS threads start/end.
Also natsThreadLocal_SetEx() in the unix implementation was not
checking the setErr that is set to false on library teardown which
was causing the library to be attempted to be opened in case of
error.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>