[ADDED] natsOptions_SetWriteDeadline()
Adds a write deadlien for all socket write operations. It means that
a publish call could fail (with NATS_TIMEOUT) if the server is not
able to consume data (either not fast enough or connection is in
a state where it is not reported as closed, but transmission is
not happening).

Resolves #217
Resolves #215

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>