[FIXED] ErrorHandler not properly invoked for SlowConsumer cases
The goal is to invoked the error handler only once for several
dropped messages, that is, until the slow consumer marker for
a subscription is cleared. There was a defect in getting to
know when to fire the callback.
Typical case would be if there is a condition that trips the
slow consumer, the callback would not be invoked until after
a new message is received (and be still over the limit).

Resolves #327

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>