[FIXED] Connections with IP ordering and deadlines
Say user defined ordering as 64 (IPv6 first, then v4), but the
server is listening only to IPv4. If the connect has a timeout
and the wait for connect ready times out (which is the case on
Windows), then there would be no chance to connect.
This PR now will split the total timeout to the number of IPs
that will be tried. This allow the above scenario to work.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>