url: fix memory leak reported by GCC 10.0.1
GCC 10.0.1, when used with '-fanalyzer', reports:

warning: leak of ‘url’ [CWE-401] [-Wanalyzer-malloc-leak]

Issue is due to an early return. The easy way to fix it, without changing
too much, is to propagate the error code. In this way the function will
take care to free data structure calling natsUrl_Destroy(url).

Signed-off-by: Paolo Teti <paolo.teti@gmail.com>