Fix GCC 10.x warning: on output truncated before terminating with nul
GCC 10.x (also on TravisCI) raise the following warning:

In file included from /usr/include/string.h:494,
             from /home/travis/build/nats-io/nats.c/src/include/n-unix.h:37,
             from /home/travis/build/nats-io/nats.c/src/natsp.h:20,
             from /home/travis/build/nats-io/nats.c/src/msg.c:14:

In function ‘strncpy’,
inlined from ‘_canonicalKey.constprop’ at /home/travis/build/nats-io/nats.c/src/msg.c:217:9:
/usr/include/x86_64-linux-gnu/bits/string_fortified.h:106:10: warning: ‘__builtin_strncpy’ output
truncated before terminating nul copying as many bytes from a string as its length [-Wstringop-truncation]
106 |   return __builtin___strncpy_chk (__dest, __src, __len, __bos (__dest));
  |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

/home/travis/build/nats-io/nats.c/src/msg.c: In function ‘_canonicalKey.constprop’:
/home/travis/build/nats-io/nats.c/src/msg.c:181:33: note: length computed here
181 |     int         lenKey  = (int) strlen(key);


This is due to the string been truncated after use strncpy().
Even if there isn't an issue in this specific case, this patch replace
strncpy() and the explicit NUL termination with strlcpy() that is the
preferred alternative to strncpy() that has a incomplete termination
semantics.

strncpy() is banned in a lots of open-source code nowadays (git is an example)

By the way strlcpy() is not available in all platforms (it is part of
muslc, but not glibc), so this patch add also a way to auto-detect if
available and adds a private implementation if not.

Signed-off-by: Paolo Teti <paolo.teti@gmail.com>