[ADDED] natsOptions_SetNKeyFromSeed() and nats_Sign()
The option provides the public key and the file containing the
seed. This will be used during connect to load the file and use
the seed to sign the nonce. The memory where the seed is copied
will be cleared.

Also exposing a way for user to sign any input given an encoded
seed.

Resolves #369

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>