[FIXED] EventLoop: possible file descriptor leak
Two issues: if using an event loop and SSL but the connection
fails before the event loop is attached, the library needs
to close the socket.
On normal socket shutdown, since there is no readloop in the
context of external event loop, ensure that the library closes
the socket properly.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>