Fix for use with retry on failed connect
When using this option and providing a "connected" callback,
the natsConnection_Connect() becomes async and we trigger the
attempts to connect using the reconnect thread. When that is
the case, no socket was opened, so exclude the "initial connect"
condition to decide if we need to wait for the socket closed
notification.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>