[FIXED] Subscription could not be closed while draining
This addresses several issues related to drain:
- Subscriptions could not be closed (unsubscribe) by the user once
draining had started
- The natsConnection_Drain() and natsSubscription_Drain() were
possibly blocking since those functions would internally call
natsConnection_Flush().

The calls are now not blocking (the flush is done in a library
thread), the subscription can be closed while drain is in progress.

Two new APIs were added:
- natsSubscription_DrainTimeout(): this allows to specify a timeout.
When the timeout is reached and if the subscription is not fully
drained, the subscription will automatically closed.
- natsSubscription_DrainCompletionStatus(): once the user has verified
that the drain is complete, this function can be used to know if the
drain completed successfully or not. It returns NATS_OK if drain
was completed without error, but would return NATS_INVALID_SUBSCRIPTION
if the subscription was closed prematurely, or NATS_CONNECTION_CLOSED
if it was the connection.

Resolves #394
Resolves #383

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>