[CHANGED] Default Ping settings
The current defaults were too small and would cause a client to
mark the connection as lost after 15 seconds or so. This means
that a server restart or leadership change may lead to the client
dropping the connection.

Made the change so that the overall time the client keeps the
connection valid is as long as the server default ping settings.

Of course, user that use stanConnOptions_SetPings() are not affected.
This is only changing the default values.
Also, in the event that the server no longer has the client, the
client will detect that at the first ping interval, which is
5 seconds (the interval has not changed).

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>