[ADDED] natsMsg_IsNoResponders API
This is used in message callbacks for subscriptions that have been
set on the reply subject for a PublishRequest calls. This will let
the application know that the request failed due to no-responders.

In other situations, such as natsConnection_Request() or using
the synchronous subscriptions, no message is returned and status
NATS_NO_RESPONDERS is returned.

This also fixes natsSubscription_NextMsg() that should have
returned NATS_NO_RESPONDERS in such situation but was not.

Resolves #421

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>