Always pass reply subject as arg to low level conn publish
Do not rely on msg->reply for requests, instead use the one
passed as the argument to low-level natsConn_publish().

This allows JS published messages to be resent easily, or
possible a JS message that is received (that would have
a JS.ACK reply to be able to be JS sent if desired).

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>