Replaces jsSub->consumer with ephemeral
We were keeping the consumer name only for the purpose of deleting
the JS consumer on Unsubscribe(), but that applies only to ephemeral
consumers. Also got rid of delCons since now `jsi->ephemeral` will
be set only when the library decides that it should delete the
consumer.

Also added check for JetStream tests to make sure that server is
at v2.3.3 or then just skip the tests.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>