[FIXED] JetStream README
The README with excerpt of APIs was done during a development phase
that preceeded a major renaming. The README.md section was unfortunately
not updated.

It is recommended to instead look at the `examples/js-xxx.c` that
are compiled as part of the build process and will give a better
idea on how to use the API and always be up-to-date

[ci skip]

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>