[CHANGED] PurgeDeletes() will now keep markers that are less than 30min old
There is a breaking change where kvStore_PurgeDeletes() accepts now
kvPurgeOptions instead of kvWatchOptions.

We needed from kvWatchOptions only the timeout, and as it standed, it was
bad since user could have passed watcher options that would have
affected the internal watcher that we make to collect delete markers.

Also, when invoking kvStore_PurgeDeletes(), the delete markers that
are older than a default of 30 minutes will be deleted, however more
recent ones will be kept. The data is always removed, even if a marker
is not.

The user can change the 30 minutes threshold using a new purge
option called kvPurgeOptions.DeleteMarkersOlderThan. If set to -1,
it restores the old behavior of deleting delete/purge markers,
regardless of their age.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>