[FIXED] JetStream: natsSubscription_Destroy should not delete JS consumer
Added comments in the js_Subscribe() calls that explains the
behavior of auto created/deleted JS consumers.
It was the intent that auto-created JS consumers would be deleted
only when explicitly calling unsubscribe/drain, similar to Go client.

However, in C, the user needs to call natsSubscription_Destroy to
free memory. This call is calling internally natsSubscription_Unsubscribe
(if the call was not made explicitly), but then we need to disable
the deletion of the JS consumer in that case.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>