[FIXED] JetStream: `jsConsumerConfig.Name` was not unmarshal'ed
The new configuration `Name` in `jsConsumerConfig` could be set
when calling `js_AddConsumer`, but when getting the `jsConsumerInfo`
response (from that call or any `js_GetConsumerInfo`, the `Name`
was present in `jsConsumerInfo`, but was not in the `Config` field
(a `jsConsumerConfig`). Although I am not sure this is needed,
making sure that we unmarshal properly.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>