[ADDED] KeyValue: Support for Mirror and Sources
Also updated state for proper Put() and Get() semantics with mirrors
and across domains, e.g. leafnodes.

In practice when a mirror is across a domain, should be named the
same as origin. That allows an app to run anywhere without anything
special in terms of domains when binding to the KV itself.

(Removed dead-code following previous PR #607)

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>