[CHANGED] Deprecate NATS_BUILD_TLS_USE_OPENSSL_1_1_API CMake variable
Building with TLS will now default to OpenSSL 1.1+/3.0+. The OpenSSL
APIs used by the C Client are the same in OpenSSL 3.0 and 1.1, so
we did not need to update code to support OpenSSL 3.0. The OpenSSL
1.1+ has an EOL of Sep 2023.

The OpenSSL 1.0.2 APIs being used can be possibly removed in a future
release. It will still be possible until then to build the NATS C
Client with OpenSSL 1.0.2 by setting NATS_BUILD_TLS_USE_OPENSSL_1_1_API
to OFF.

Resolves #614

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>