[CHANGED] Default connection error handler.
If no error handler is specified by the user, the library will
use a default error handler that prints some error information
to stderr.

If the user did not specify an error handler but does not wish
to have anything printed to stderr, the user should set a dummy
error handler that does not do anything.

Resolves #622

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>