Fixed download of NATS Server for latest/main (#711)
The "latest" was not working and would always use the one from
nats.c.deps.
As for "main", I noticed lately that sometimes, we would get
a binary of size 0, which would make all tests using the nats-server
to fail. So make sure that we can run the server and get a valid
version string. If not, repeat the test up to 5 times, otherwise
fail the run.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>
Co-authored-by: Lev <1187448+levb@users.noreply.github.com>