[FIXED] Build: issue on Android with NATS_EXTRA_LIB
A user (@psantiagol) reported an issue (in #64) building on Android
due to the use of `NATS_EXTRA_LIB`. This PR skip the definition of
this variable if building on Android. Should other users then report
that this is breaking their build on Android, we will have to make
this a conditional based on a boolean build option.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>