[FIXED] Build: issue on Android with `NATS_EXTRA_LIB` (#739)
* [FIXED] Build: issue on Android with NATS_EXTRA_LIB

A user (@psantiagol) reported an issue (in #64) building on Android
due to the use of `NATS_EXTRA_LIB`. This PR skip the definition of
this variable if building on Android. Should other users then report
that this is breaking their build on Android, we will have to make
this a conditional based on a boolean build option.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>

* Fix test JetStreamBackOffRedeliveries

With PR https://github.com/nats-io/nats-server/pull/5242 in server
code, not providing a MaxDeliver is considered unlimited and should
not error out. So fixing test to set MaxDeliver <= to the backoff
list to get the expected error.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>

---------

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>