[ADDED] TLS: natsOptions_TLSHandshakeFirst() (#780)
* [ADDED] TLS: natsOptions_TLSHandshakeFirst()

This is to force a client to perform the TLS handshake first, that
is, not wait for the INFO protocol from the server. This is needed
if the server is configured to require clients to perform the TLS
handshake first (before sending the INFO protocol).

Resolves #779

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>

* Fixed flapper

If the cluster was not fully established when the rest of the test
was running, the library may get additional client URLs to reconnect
to (gossip protocol). Looks like this was happening more on Windows CI
than Linux.

Using some options to clamp all that.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>

---------

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>