[FIXED] Added missing NATS_EXTERN on some functions.
Namely `nats_OpenWithConfig` and `natsConnection_Reconnect`.

Also normalized the `NATS_EXTER <return type>` in one line and
the function name on the next line.

Updated the C++ compile test since it was failing on my mac env,
and modified `nats_StringEquals` that was triggering a warning
on my Windows VM build.

Resolves #835

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>