Fixed compiler issue on macOS related to assert for status text
Made this a test instead. Also fixed a flapper that was due to a
recent change in the NATS Server where an INFO in response of the
CONNECT protocol is now sent, causing the override server version
that we use in the test to be replaced with the actual server
version.

Relates to #873

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>