[FIXED] Build: Unknown type name SSL_verify_cb
We had several issues around the introducing of a verification callback.
We reverted to use a `NATS_WITH_EXPERIMENTAL` build flag, but still,
as part of the rewrite, there was something of the like:
```
SSL_verify_cb cb = _collectSSLErr;
```
Prior to 3.10, we would have called `SSL_set_verify(ssl, SSL_VERIFY_PEER, _collectSSLErr);`
which was not causing an issue. The above would when building with
OpenSSL prior to 1.1.1.

So move the `cb` initialization as a `SSL_verify_cb` only when under
the experimental flag, otherwise, invoke the same code we had prior
to 3.10.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>