[FXIED] LibEvent: natsConnection_Close() not closing the TCP connection
This was likely introduced in PR #815 when delaying the close of the
socket based on the event loop adapter invoking a NATS C callback
to indicate that it is safe to do so. The `libuv` adapter is doing
the right thing, however the `libevent` was not since it was creating
a one-time event based on `EV_TIMEOUT` passing the socket to close,
however, in that mode the socket from the callback is always set
to `-1`. We have to pass the socket to close though the `nle` object.

I also made the `natsLibevent_Detach()` function schedule an event
to ensure that a `_freeCb` is now invoked *after* execution of the
`_closeCb`.

Resolves #879

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>