[FIXED] EventLoop: Libuv may crash if connection destroyed while consuming
If the user destroys the connection while the read event was in progress
it could be possible for the library to crash trying to process a read
event.

We need to immediately stop polling when the read event is removed.

Changing also the way we retain the connection: we will now retain
on the very first attach, and the adapters will have to call
`natsConnection_Destroy()` when the event loop is fully detached.

Resolves #888

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>