[FIXED] MicroServices: Wrong marshaling of `average_processing_time`
The field was incorrectly marshaled as a string duration instead
of a number (nano seconds). Also, the computation of the average
time would not take into account the possibility of having zero
number of requests, which would produced a NaN and an invalid
number during marshal.

Added test to check statistics before any request and that the
field is a proper number (not a string). Also verified that it
now works well with the NATS CLI.

Resolves #890

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>