[FIXED] TLS: Possible hang during handshake
The reason was that we switched the connection to blocking mode
for the SSL handshake duration. In some cases, this could lead to
the connection being blocked in the handshake for too long (longer
than the provided connection timeout).

We now don't switch to blocking mode and check if `SSL_do_handshake`
wants a read or write and wait for such event to be available.

Resolves #906

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>