[CHANGED] TLS: Changed `SSL_verify_cb` to `natsSSLVerifyCb`
The callback uses a `void*` that the user should cast to a `X509_STORE_CTX*`.

This removes the `nats.h` dependency on `NATS_HAS_TLS` and openssl headers.

Programs that have a verification callback would have to modify it
to change to a `void*` and do a cast, as described in the doc for
the `natsSSLVerifyCb` verification callback.

I have tested that an application compiled dynamically (with 3.10.1)
and code that directly uses the `X509_STORE_CTX*` would work without
problems if linked to a library build with this code.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>