[ADDED] Option `natsOptions_AllowConcurrentTLSHandshakes`
We used to have a ssl context lock held when setting up SSL when
connecting, which included the SSL handshake.

Introduce an option that allows the user to explicitly allow
the connections created with the same `natsOptions` object
to perform the SSL handshake concurrently. If any SSL callback
is configured, the user must ensure that their implementation
is thread-safe, or not use the new option.

Resolves #899

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>