[IMPROVED] TLS: Speed up concurrent connects
We used to have a ssl context lock held when setting up SSL when
connecting, which included the SSL handshake.

There was no real reason to hold that lock because even if all
connections shared the same `natsOptions` (which holds the ssl
context), the ssl context fileds are immutable. If the user were
to invoke any `natsOptions` setter that would affect the context,
a new context was created. See `_getSSLCtx` for details.

In this PR, I have also fixed a test introduced in the previous
PR because it could have failed in environment that are very slow
with a combination of running with valgrind.

Resolves #899

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>