[FIXED] Options: don't allow modifications if SSL configured and in use.
If a `natsOptions` object is configured with TLS/SSL, once it is
passed to a `natsConnection_Connect` call, the options cannot be
modified because internally we use a `SSL_CTX` that cannot be
modified once `SSL` object are created from it.

We used to detect that the options were in use but instead of
returning an error, we would dissociate the context from connections
and create a new `SSL_CTX` for this `natsOptions` object. It "worked"
in that connections were safely not using a context that could change,
but it was wrong because the `SSL_CTX` would obviously be created
empty and not reflect the previous modifications that were made.

It is very unlikely that users would create a `natsOptions`, pass
it to connections and then modify it, but this PR makes it explicit
that using it this way is wrong and would lead to unexpected behavior.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>