[FIXED] Connection drain could cause missed reply and/or 100ms delay
The more important issue is that we should have drained the internal
`respMux` subscription after all user subscriptions have been drained
so that we don't miss replies that would be sent by the user subscriptions
callbacks.

Draining the internal `respMux` subscription would often cause the
check for number of subscriptions to be positive, which then would
lead to a wait of 100ms before checking again that the number
would be down to 0.

This PR addresses both issue: make sure that the internal subscription
is drained last, and use condition variable to notify when subscriptions
are removed and the number is down to the expected number.

Resolves #911

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>