[FIXED] EventLoop: Handling of possible failure on initial attach
On initial attach, the internals needs to be free, but we can't
call the "detach" function neither on initial nor successive
attach.

Introduced `natsConnection_ProcessDetachedEvent` function that
is invoked by the adapters when fully detached. This allows the
library to release resources associated with the connection after
it has been detached.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>