[ADDED] `natsConnection_GetName` to return the name of the connection.
A connection name can be set with `natsOptions_SetName()` and is
sent to the server as part of the `CONNECT` protocol.

This new API allows applications to know the name that was
set for a connection object.

The returned string must be freed by the user when no longer needed.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>