[FIXED] Handling of mutex creation failures
In some cases, we were creating the mutex as part of an object
initialization and at the end, if a failure occurred we would call
the "release" API to destroy that object. This would not be good
because if the mutex creation itself failed, the "release" function
would crash. So call the "free" function instead of the "release".
In some cases, it is required to call the "release" because that
also takes care of stopping/undoing this that have been done
during the object creation. In those cases, make sure that the
mutex creation is done "first" and handle specifically the creation
failure.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>